Drastically simplify builtins
We create a way to turn curried Expr functions to a function that
takes a list of Exprs. This allows us to take all of the functions
that eval all of their arguments and put them into a keyword map
instead of manually pattern-matching on each of them like a pleb.