Add stateful evaluation
This is absolutely terrible, and I'm sure that Katie is going to
look at this code and cringe. All sorts of monadic operators and
various lifting thingies could be used here, but I have absolutely
no idea what I'm doing.

This is going to get further complicated when I add lexical closures,
as the state is going to contain *two* things - the current local
variables captured by lambda function declarations, and the global
variables that `set!` impacts.